requirements.txt
asgiref==3.8.1
autopep8==2.3.1
certifi==2024.7.4
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==2.2.0
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.8
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
psycopg2-binary==2.9.9
pycodestyle==2.12.1
python-dotenv==1.0.1
requests==2.32.3
setuptools==73.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
